Move classes into proper packages
